Break out `Cargotest` into its own `cargotest` module
